Improved the Undo mechanism
Description:
The problem was in Event registration - the constructor in Abstract class had registration command -> an user could made command and called it twice. In the result - command was repeated, but manager knowed about only first call.

Solution:
Don't use the AbstractCommand's constructor for registering a command -> let's take a state of CommandManager and registration if it's possible